conditional import of matcher
