+++ internal/scripts/buildts/main.go
@@ -49,7 +49,9 @@ func buildKit() {
 			"./kit/_typescript/fmt/fmt.ts",
 			"./kit/_typescript/json/json.ts",
 			"./kit/_typescript/listeners/listeners.ts",
-			"./kit/_typescript/matcher/matcher.ts",
+			"./kit/_typescript/matcher/register.ts",
+			"./kit/_typescript/matcher/find_best_match.ts",
+			"./kit/_typescript/matcher/find_nested_matches.ts",
 			"./kit/_typescript/theme/theme.ts",
 			"./kit/_typescript/url/url.ts",
 		},
