--- kit/_typescript/matcher/README.md
@@ -1,2 +1,0 @@
-This is an unpublished experiment that ports the Go matcher pkg to TypeScript.
-It's basically a line-for-line port and includes benchmarks to compare perf.
