+++ package.json
@@ -54,9 +54,17 @@
 			"import": "./npm_dist/kit/_typescript/listeners/listeners.js",
 			"types": "./npm_dist/kit/_typescript/listeners/listeners.d.ts"
 		},
-		"./kit/matcher": {
-			"import": "./npm_dist/kit/_typescript/matcher/matcher.js",
-			"types": "./npm_dist/kit/_typescript/matcher/matcher.d.ts"
+		"./kit/matcher/register": {
+			"import": "./npm_dist/kit/_typescript/matcher/register.js",
+			"types": "./npm_dist/kit/_typescript/matcher/register.d.ts"
+		},
+		"./kit/matcher/find-best": {
+			"import": "./npm_dist/kit/_typescript/matcher/find_best_match.js",
+			"types": "./npm_dist/kit/_typescript/matcher/find_best_match.d.ts"
+		},
+		"./kit/matcher/find-nested": {
+			"import": "./npm_dist/kit/_typescript/matcher/find_nested_matches.js",
+			"types": "./npm_dist/kit/_typescript/matcher/find_nested_matches.d.ts"
 		},
 		"./kit/theme": {
 			"import": "./npm_dist/kit/_typescript/theme/theme.js",
